requirements.txt
Django==4.0.4
djangorestframework==3.14.0
docker==6.0.1
celery==5.2.7
redis==4.3.3
flower==1.0.0
djoser==2.1.0